requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
typing_extensions==4.14.1

psycopg
django-environ

djangorestframework
djangorestframework-gis

Pillow

requests
stac_validator
jsonschema>=4.0.0
